refactor: move data loading into conftest
